Move all benchmark tests to benches folder
This allows ignoring benches on stable
