+++ src/data_structures/suffix_array.rs
@@ -459,8 +459,6 @@ impl PosTypes {
 
 #[cfg(test)]
 mod tests {
-    use test::Bencher;
-
     use super::*;
     use super::{PosTypes,SAIS,transform_text};
     use bit_vec::BitVec;
@@ -542,11 +540,5 @@ mod tests {
         let text = b"TGTGTGTG$";
         let pos = suffix_array(text);
         assert_eq!(pos, [8, 7, 5, 3, 1, 6, 4, 2, 0]);
-    }
-
-
-    #[bench]
-    fn bench_suffix_array(b: &mut Bencher) {
-        b.iter(|| suffix_array(b"GCCTTAACATTATTACGCCTA$"));
     }
 }
